feat, refactor(TextInput): 음식점 추가에 사용되는 인풋 분리 및 글자수 표시 기능 구현
